backend/src/requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.3
django-crispy-forms==1.7.0
django-filter==1.1.0
django-guardian==1.4.9
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.14.2
openpyxl==2.5.1
pandas==0.22.0
PyJWT==1.5.3
python-dateutil==2.7.0
pytz==2017.3
requests==2.18.4
rest-pandas==1.0.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
